requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.58.2
django-appconf==1.0.6
django-widget-tweaks==1.5.0
docopt==0.6.2
idna==3.6
oauthlib==3.2.2
Pillow==10.1.0
pipreqs==0.4.13
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
sqlparse==0.4.4
urllib3==2.1.0
whitenoise==6.6.0
yarg==0.1.9
